Show progress indicators for corpus entries only while they are in progress.
This change reduces the number of worker threads to equal the number of promoted sources (4).
I tested this out with 8+ canned sources that delayed for a few seconds to see how
it would look, and it is easier to understand what is going on when the progress
indicator next to each corpus entry only starts when it is actually working.

One other tweak is that we don't show the number of results in the description until
the source has finished.