Don't show corpus entries that have reported zero results unless they have been viewed before.
The rules for showing a corpus result when the "more results" is expanded is now:
- they haven't reported yet
- they have results
- they have been shown before

Thus, if a source reports zero results before the user ever expands the "more results",
we don't show an entry for it once the user does expand the results.

Also includes a fix for deciding whether a cached result is promoted.