Add permissions for talking to GlobalSearch.
This fixes two security issues in GlobalSearch:

http://b/issue?id=1871088 (Apps could use GlobalSearch to read
content providers that they do not have permission to access)

http://b/issue?id=1819627 (Apps could launch any intent by sending
it to GlobalSearch)

The permission are declared "signature", so only system apps should
have them. The search dialog now runs in the system process instead
of each applicaiton process, so it will be able to access GlobalSearch.