Rank sources based on CTR if they have enough impressions, give others some exposure until they do.
The new requirement was to have new sources be gauranteed to not show up in the promoted list right away.
But we'd also like to give them some exposure (e.g not dumped to the bottom of the list).  This change
acheives both of those while still preserving the goal of not having a source with a high CTR on few
impressions get ranked highly.

Now the ranking is:
- web source
- rest of promoted source slots used by highest CTR sources
- unranked sources
- rest of ranked sources

This means we no longer use priors in the shortcut repository, and instead only rank sources with at least
X=5 impressions.  The logic of mixing in the web / ranked / unranked is elsewhere.