Fix synchronization bug in SuggestionSources.
Before, the Settings.System#SHOW_WEB_SUGGESTIONS observer called
updateEnabledSources(), which is not synchronized.
Instead it now calls updateSources(), which is.
This will do a bit more work, but it's consistent with how
the ACTION_SETTINGS_CHANGED broadcast intent is handled,
and it's only done when the search settings are changed,
which should be rare.