Be as lazy as possible when construction the SuggestionCursor, improving responsiveness (mainly by avoiding GCs).
If we believe the user is still typing, instead of just delaying sending
off the queries, we avoid doing almost everything until timeout has expired
without being cancelled.  SuggestionCursor is now created without a backer
initially.  The backer is attached once it is ready.

This seems to make typing much more responsive, mainly because we avoid
allocating memory in the lazy case, and therefore avoid GCs that were the
cause of the hiccups.