Display suggestions from Music app properly.
The music suggestions provider (i.e. media provider) now provides icons & text
usable in global search so switch to using that. In the process fix a bug in
global search where we were assuming suggestions come from the same activity
which had the searchable info, which was not true in the Music app's case.

Related bug - http://b/issue?id=1819601