Include intent extra data in shortcut refresh request
The value of the suggest_intent_extra_data column of a shortcut
is now passed as a query parameter (with key suggest_intent_extra_data)
in the shortcut refresh URI.

This makes it possible for sugggestion sources to get access to the
full suggestion data when refreshing, if the source encodes it
as a string and puts it in the suggest_intent_extra_data column.
This is for example useful if the server request when refreshing
oneboxes and business listings fails.