Correctly handle shortcut refreshing for isSpinnerWhileRefreshing.
We avoid caching the entire refresh; only if isSpinnerWhileRefreshing is true,
we store the updated icon2 so we can correctly remove the spinner when we aren't
going to make the refresh query.

Also, the key for the cache is based on the component and shortcut key now insteaad
of just the shortcut key.

Fixes http://b/issue?id=1992533 and http://b/issue?id=1996220