Remove "Show web suggestions" setting from GlobalSearch
Each web search engine will instead have a setting UI for this.
The setting is stored in the system settings, but the UI
is now under each engine. This is to allow engines to disable
the setting if they don't support suggestions, and to make the
settings more coherent. For example, EnhancedGoogleSearch
will how have the "Show web suggestions" and "Show search history"
settings together, with the latter depending on the former.

Also, to support one settings activity(-alias) for each search engine
in a package that contains multiple engines, we first look for
an exact component name match when looking for a settings activity.

This is the GlobalSearch part of http://b/issue?id=1996992