Only record impressions after a click
To avoid recording impressions for suggestions that are only
display for a short time while the user is typing, we now
only record impressions once a suggestion has been clicked.
This should also improve typing performance a bit.

This change also does some refactoring of SuggestionSession.

This is the GlobalSearch part of http://b/issue?id=2000655