Demote Music source by default.
This is done by adding a new category of unpromotable sources.
The music source is the only unpromotable source, and
it is only made unpromotable if it is not ranked high enough
to make it into the top NUM_PROMOTED_SOURCES.

This change should have the effect
of demoting Music as long as it	doesn't	have enough CTR	to
be promoted, and as long as the	user doesn't turn off
one of the other sources.

TODO: Once the Music app returns better suggestions (i.e. token
prefix matches, rather than string infix matches) we should make it
promotable again.

Fixes http://b/issue?id=2018021