Added external logging to GlobalSearch
GlobalSearch now logs clicks to an external content
provider. Only clicks that come from the application
that contains the click log provider are logged. The
log requests include the position of the clicked
suggestion, and the types of suggestions in the other
slots. Currently the only types used are WEB,
BUILT-IN (i.e. GlobalSearch suggestions such as "Search the web"
and "Go to website"), and OTHER (Contacts, Apps, Music, 3rd parties).

The click log provider is protected by a permission to avoid log stuffing,
and the click log provider must have a special permission to
avoid log stealing.

This is part of the fix for http://b/issue?id=2097469

Change-Id: I399202ba67db5eb728b38b8135c59c5f3cdf1b58