Use a predicatable synchronous executor in tests.
Some of the PerTagExecutorTest tests were failing. This seems to
be because of changes to thread scheduling which mean that the old 1 second
timeout was not enough to ensure that the test threads always had
time to finish.

Should fix:

http://b/issue?id=2117625
com.android.globalsearch.PerTagExecutorTest:testPendingRuns_intermediateDropped is failing

http://b/issue?id=2115644
com.android.globalsearch.PerTagExecutorTest:testPendingRuns is failing

Change-Id: I4bba9780f822efd529d2803dbe9eb2aee523db01